auroraml-movie-recommendations/requirements.txt
streamlit==1.37.0
python-dotenv
psycopg2-binary
pgvector
urllib3==1.26.6
boto3>=1.28.57
awscli>=1.29.57
botocore>=1.31.57
anthropic==0.43.1
